requirements.txt
arrow==1.3.0
asgiref==3.8.1
backoff==2.2.1
bcrypt==4.2.1
build==1.2.2.post1
cachetools==5.5.0
chroma-hnswlib==0.7.6
chromadb==0.6.3
coloredlogs==15.0.1
Deprecated==1.2.15
durationpy==0.9
fastapi==0.115.6
filelock==3.16.1
flatbuffers==24.12.23
fqdn==1.5.1
fsspec==2024.12.0
google-auth==2.37.0
googleapis-common-protos==1.66.0
grpcio==1.69.0
httptools==0.6.4
huggingface-hub==0.27.1
humanfriendly==10.0
importlib_metadata==8.5.0
importlib_resources==6.5.2
isoduration==20.11.0
jsonpointer==2.1
jupyter==1.1.1
jupyterlab==4.3.4
kubernetes==31.0.0
langchain-core==0.3.30
langchain-openai==0.3.1
markdown-it-py==3.0.0
mdurl==0.1.2
mkl-service==2.4.0
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
notebook==7.3.2
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.59.9
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
ply==3.11
posthog==3.8.3
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyPika==0.48.9
pyproject_hooks==1.2.0
PyQt5==5.15.10
regex==2024.11.6
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
shellingham==1.5.4
starlette==0.41.3
sympy==1.13.3
tiktoken==0.8.0
tokenizers==0.21.0
typer==0.15.1
types-python-dateutil==2.9.0.20241206
uri-template==1.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
webcolors==24.11.1
webencodings==0.5.1
websockets==14.2
wrapt==1.17.2
zipp==3.21.0
